Group phpunit tests by functionality tested
