+++ tests/phpunit/Performance/CreateTaxonomyTermTest.php
@@ -1,17 +1,16 @@
 <?php
 
-namespace tests\phpunit;
+namespace tests\phpunit\Performance;
 
 use weitzman\DrupalTestTraits\ExistingSiteBase;
-use Drupal\media\Entity\Media;
 
 /**
- * A test to confirm ability to create media.
+ * A test to confirm ability to create taxonomy.
  */
-class CreateMediaTest extends ExistingSiteBase {
+class CreateTaxonomyTermTest extends ExistingSiteBase {
 
   /**
-   * A test method to determine the ability and time to create media.
+   * A test method to determine the ability and time to create a node.
    *
    * @group performance
    * @group functional
@@ -19,7 +18,8 @@ class CreateMediaTest extends ExistingSiteBase {
    *
    * @dataProvider benchmarkTime
    */
-  public function testCreateMedia($benchmark) {
+  public function testCreateTaxonomyTerm($benchmark) {
+    $vocab = \Drupal::entityTypeManager()->getStorage('taxonomy_vocabulary')->load('health_care_service_taxonomy');
 
     // Start timer.
     $mtime = microtime();
@@ -27,25 +27,9 @@ class CreateMediaTest extends ExistingSiteBase {
     $mtime = $mtime[1] + $mtime[0];
     $starttime = $mtime;
 
-    // Creates media. Will be automatically cleaned up at the end of the test.
-    $file = file_save_data(uniqid(), 'public://test' . uniqid() . '.txt');
-
-    $media_document = Media::create([
-      'bundle' => 'document',
-      'name' => 'test',
-      'uid' => '1',
-      'field_media_file' => [
-        'target_id' => $file->id(),
-      ],
-    ]);
-    $media_document->setPublished(TRUE)
-      ->save();
-
-    $media_document->setName(uniqid())
-      ->setPublished(TRUE)
-      ->save();
-
-    $this->assertNotEmpty($media_document->getName(), 'Failed to create media');
+    // Creates a term. Will be automatically cleaned up at the end of the test.
+    $term = $this->createTerm($vocab);
+    $this->assertNotEmpty($term->getName(), 'Failed to create term');
 
     // End timer.
     $mtime = microtime();
@@ -53,9 +37,6 @@ class CreateMediaTest extends ExistingSiteBase {
     $mtime = $mtime[1] + $mtime[0];
     $endtime = $mtime;
     $microsecs = ($endtime - $starttime);
-
-    // Cleanup the file.
-    $file->delete();
 
     // Test assertion.
     $secs = number_format($microsecs, 3);
